Moved S3 support into a subpackage.
