+++ src/main/resources/io/jenkins/plugins/artifact_manager_jclouds/s3/S3BlobStore/config.properties
