--- src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-deleteArtifacts.html
@@ -1,3 +1,0 @@
-<div>This parameter is set by the property <code>-Dio.jenkins.plugins.artifact_manager_s3.S3BlobStoreConfig.deleteArtifacts</code>
-    ,if not checked the artifacts will not be deleted from S3 when the corresponding build is deleted.
-</div>
