--- src/main/resources/io/jenkins/plugins/artifact_manager_s3/S3BlobStoreConfig/help-deleteStashes.html
@@ -1,3 +1,0 @@
-<div>This parameter is set by the property <code>-Dio.jenkins.plugins.artifact_manager_s3.S3BlobStoreConfig.deleteStashes</code>
-    ,if not checked the stash will not be deleted from S3 when the corresponding build is deleted.
-</div>
